refactor(hipsr): move bufferization models out of the public header
## Summary

The hipsr `BufferizableOpInterface` models move from a public header to
`lib/Dialect/Hipsr/Transforms/BufferizableOpInterfaceImpl.cpp`, behind a
single registration entry point. `hipsr.preserve_shape` is still the only
op with a model and its bufferized output is unchanged.

## Why

`include/hip/Dialect/Hipsr/IR/HipsrBufferize.h` defined
`PreserveShapeBufferizableModel` inline in a public header that also
pulled in `HipsrOps.h` and `BufferizableOpInterface.h`. Every translation
unit that included it recompiled the model bodies, and the model name was
publicly reachable, so a caller could attach a partial set of models.

Upstream MLIR keeps these in
`<Dialect>/Transforms/BufferizableOpInterfaceImpl.cpp` in all nine
dialects that implement the interface. No upstream dialect has a
`Bufferization/` subdirectory.

## What

- The model moves into an anonymous namespace nested in `mlir::hipsr`, so
  it now has internal linkage and only the registration function can
  attach it.
- `registerBufferizableOpInterfaceExternalModels` replaces
  `registerHipsrBufferizableOpInterfaceModels`, matching the eight
  upstream calls it sits beside in `InitAllPasses.h`. It is declared in a
  header that has no includes.
- `bufferize` uses upstream's `replaceOpWithNewBufferizedOp` instead of
  hand-rolling create-then-erase.
- The early exit for an already-memref `$data` is removed. The
  bufferization driver gates every `bufferize` call on
  `hasTensorSemantics`, and `hipsr.preserve_shape` has tensor semantics
  only when `$data` is a tensor, so the branch was unreachable.
- The LIT test moves to
  `test/lit/Dialect/Hipsr/Transforms/OneShotBufferize/preserve_shape.mlir`
  with its contents unchanged, matching the PascalCase pass-named
  siblings.

## Test plan

Local incremental build, the `hipsr`-filtered LIT run (26 passed, 0
failed) and the full LIT suite, plus `pre-commit run --all-files`. The
relocated test passing with untouched RUN and CHECK lines is the evidence
that the move is behavior-neutral.

`Conversion/onnx-to-hip/test_gather_block_quantized.mlir` fails in the
full suite. It is unrelated and pre-existing: it asserts inside
`IntegerAttr::getSInt` during `--convert-onnx-to-hip`, and it reproduces
on `main` with these changes stashed.

## Notes for reviewers

The generic DPS model and the trait-filtered bulk attach are left out on
purpose. hipsr DPS ops constrain operands to `Hipsr_TensorOrDeviceMemRef`
while stock One-Shot Bufferize produces space-less memrefs, so those
models cannot be tested until the memory-space question is settled.

## AI assistance

Cursor (Claude Opus 5) assisted with the relocation, the upstream-helper
cleanup, and the test move. Validated with a local incremental build, the
MLIR LIT suite, and `pre-commit run --all-files`.

Made-with: Cursor
