+++ include/hip/InitAllPasses.h
@@ -10,8 +10,8 @@
 #include "hip/Compiler/PluginRegistry.h"
 #include "hip/Conversion/OnnxToHip/Passes.h"
 #include "hip/Conversion/Passes.h"
-#include "hip/Dialect/Hipsr/IR/HipsrBufferize.h"
 #include "hip/Dialect/Hipsr/IR/HipsrDialect.h"
+#include "hip/Dialect/Hipsr/Transforms/BufferizableOpInterfaceImpl.h"
 #include "hip/Dialect/Hipsr/Transforms/Passes.h"
 #include "hip/Dialect/IR/HipBufferize.h"
 #include "hip/Dialect/IR/HipDialect.h"
@@ -89,7 +89,7 @@ inline void registerAllDialects(mlir::DialectRegistry &registry) {
       registry);
   mlir::memref::registerAllocationOpInterfaceExternalModels(registry);
   mlir::hip::registerHipBufferizableOpInterfaceModels(registry);
-  mlir::hipsr::registerHipsrBufferizableOpInterfaceModels(registry);
+  mlir::hipsr::registerBufferizableOpInterfaceExternalModels(registry);
 }
 
 /// Load all required dialects into an MLIRContext.
