+++ lib/Dialect/Hipsr/Transforms/CMakeLists.txt
@@ -4,6 +4,7 @@
 ##
 
 add_library(HipsrDialectTransforms STATIC
+  BufferizableOpInterfaceImpl.cpp
   ExternalizeConstants.cpp
   HipsrPoolAllocPass.cpp
   MaterializeInitTensorsPass.cpp
@@ -19,6 +20,7 @@ add_dependencies(HipsrDialectTransforms
 target_link_libraries(HipsrDialectTransforms PUBLIC
   HipsrDialectIR
   HipConstantsIO  # writeConstantsBinToFileSystem / ConstantEntry
+  MLIRBufferizationDialect  # BufferizableOpInterface models
   MLIRPass
   MLIRTransformUtils
   MLIRFuncDialect
